move textpointercursor to separate file
