+++ Runtime/Behaviors/Cursors/TextPointerCursor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 49e6f03934554489aa5fb2faea24b791
+timeCreated: 1727689747
